+++ lwk_bindings/src/tx_builder.rs
@@ -80,7 +80,7 @@ impl TxBuilder {
     pub fn drain_lbtc_to(&self, address: &Address) -> Result<(), LwkError> {
         let mut lock = self.inner.lock()?;
         let inner = lock.take().ok_or(LwkError::ObjectConsumed)?;
-        *lock = Some(inner.drain_lbtc_to(address.into())?);
+        *lock = Some(inner.drain_lbtc_to(&address.into())?);
         Ok(())
     }
 
@@ -88,7 +88,7 @@ impl TxBuilder {
     pub fn drain_lbtc_to_explicit(&self, address: &Address) -> Result<(), LwkError> {
         let mut lock = self.inner.lock()?;
         let inner = lock.take().ok_or(LwkError::ObjectConsumed)?;
-        *lock = Some(inner.drain_lbtc_to_explicit(address.into())?);
+        *lock = Some(inner.drain_lbtc_to_explicit(&address.into())?);
         Ok(())
     }